GUI: Render to highest precision
For delegates panel, we show 0 for < 1 sem
As we allow fractional SEM votes, we should show
to same precison as elsewhere.

Also only show precision to just number available.
So we will render 3 instead of 3.000 and 3.2 instead
of 3.200.  This can allow us to choose to render more
decimals if we want without cluttering UI if user only
uses full SEM.