GUI: Don't update dropdowns if user is interacting
If we update select boxes while user is interacting with it,
It re-renders the page and loses context/state of scrolling.

This is annoying for users.  If they are interacting with
the box, we have high confidence that data is not changing
elsewhere, or it can wait til next refresh.