Consensus: add method 'validateSetHashes' to fastSync
The method 'validateSetHashes' was added to fastSync for better code readability.

This method now also validates the block header, in addition to the hash validation.

Finally, it no longer compares the block number against latest block number as blocks are added to currentSet only if block number is greater than latest.