Net: Inform user of failure to sync due to Peer exception
We have many users that are unable to determine why they are unable to
sync.  It is often due to using a proxy/VPN, and some oddities in
which IP address to use.

I feel it is appropriate to at least give users a clear message as to
why the failure is occurring.

This can be replaced in the future with a better lookup mechanism.
I am open to alternatives.  Currently we do not have granularity to
determine why the peer is invalid, but we may want to give a specific
error code, rather than a general 'invalidpeer' message on denial.