VM: Stub implementation of VM client classes
This is a work in progress, mostly to see what changes are required to
semux core to support the eth VM.

These classes should eventually be removed and implement these interfaces
natively.  But to keep scope of work down or in central location,
facade classes seemed cleaner to see what core changes are necessary.