Util: Update to capture and log OSHI errors
We have users that have had problem with OSHI errors causing
semux to be unable to start.

This is not a fatal error, so catch exceptions and log them, but still
permit the wallet to start.

Other PR has updated oshi version to hopefully catch root cause, but
future windows updates can cause similar problems