Concensus: For choosing sync height, we do want to trust the lower of the threshold.
If there are 10 validators, with heights of
10
20
30
40
etc

66% floored gives us taking a value of 60, whereas
66% rounded gives us sync height of 70

In one case, we have 5/10 that are sync'd higher, and in the
other only 4/10 that are syncd to that height or higher.

Ideally, we'd probably want the inverse, so 66% are synced higher, but I
imagine there's room to allow the 66th percent to sync, so sticking
with pre-existing logic for sync.