Core: if original validators are unable to connect, do not deny sync
In order to have full decentralization, we cannot require that the
very first validators (semux1-4) are required for syncing.

Also in order to scale past a certain level, the validators cannot be
connected to from every client.

Currently, if semux1-4 go down (Even though they are no longer valiators)
clients are unable to sync from block 0.

This change allows for choosing to sync height based on active peers
if UR validators are unable to be found.