[WIP] Core - Modularize Core
Start to break apart the project.  the API can be utilized outside project
as well as client

The core/server code can be broken apart from swing GUI, as it will be the
same regardless of UI implementation.

The UI currently uses the server for data, but could be abstracted to interact
with either a local server, or an API gateway for calls.

Start to break functional pieces into modules.

This is a WIP for now